refactor(cli): extract dump-types into its own module
Move all dump-types machinery out of main.rs into crates/ry-cli/src/
dump.rs: the TypesDump/FileDump/ScopeDump/BindingDump shapes, the
--position clap value parser, offset/position conversion helpers, the
local-binding scan and scope-indexing walkers, assemble_file_dump,
the enclosing-scope helpers, run_dump_types, and their unit tests.

The move is mechanical: run_dump_types keeps calling the existing
shared helpers (config discovery, discover_r_files, report_truncation,
sort_and_deduplicate_paths, load_user_stubs, read_r_source,
enclosing_package_root) and check.rs entry points
(CheckInput, check_project_with_scope_capture); only pub(crate)
visibility was added where main.rs dispatch needs it.

main.rs shrinks from 2772 to 1947 lines (roughly -825). Behavior is
unchanged.
