+++ crates/ry-cli/src/dump.rs
@@ -1,0 +1,836 @@
+//! `ry dump-types`: one analysis pass over the requested files, then a
+//! JSON dump of every recorded lexical scope (see ry_checker::ScopeRecord).
+//! The dump reuses `ry check`'s pipeline end to end — config discovery,
+//! file discovery, per-package grouping, workspace resolution — so the
+//! emitted types are exactly what a `ry check` run infers.
+
+use std::collections::HashMap;
+use std::path::PathBuf;
+use std::process::ExitCode;
+
+use miette::{IntoDiagnostic, Result};
+
+use ry_config as config;
+
+use crate::check;
+use crate::{
+    enclosing_package_root, load_user_stubs, read_r_source, report_truncation,
+    sort_and_deduplicate_paths,
+};
+
+#[derive(serde::Serialize)]
+struct TypesDump {
+    files: Vec<FileDump>,
+}
+
+#[derive(serde::Serialize)]
+struct FileDump {
+    path: String,
+    scopes: Vec<ScopeDump>,
+}
+
+#[derive(serde::Serialize)]
+struct ScopeDump {
+    kind: &'static str,
+    name: Option<String>,
+    start: (usize, usize),
+    end: (usize, usize),
+    bindings: Vec<BindingDump>,
+}
+
+#[derive(serde::Serialize)]
+struct BindingDump {
+    name: String,
+    kind: &'static str,
+    #[serde(rename = "type")]
+    type_: String,
+    start: Option<(usize, usize)>,
+}
+
+/// clap value parser for `--position LINE:COL`. Rows and columns are
+/// 1-based, matching the dump output.
+pub(crate) fn parse_dump_position(value: &str) -> Result<(usize, usize), String> {
+    let (line, col) = value
+        .split_once(':')
+        .ok_or_else(|| format!("expected LINE:COL, got `{value}`"))?;
+    let line = line
+        .trim()
+        .parse::<usize>()
+        .map_err(|_| format!("invalid line in `{value}`"))?;
+    let col = col
+        .trim()
+        .parse::<usize>()
+        .map_err(|_| format!("invalid column in `{value}`"))?;
+    if line == 0 || col == 0 {
+        return Err(format!("positions are 1-based, got `{value}`"));
+    }
+    Ok((line, col))
+}
+
+/// The type string for one binding. Same `Display` rendering the LSP
+/// inlay hints show, except the fully-uninformed type is reported as
+/// "unknown" so consumers never mistake `opaque<len=?>:?` for a real
+/// inference result.
+fn dump_type_string(t: &ry_core::RType) -> String {
+    if *t == ry_core::RType::unknown() {
+        "unknown".to_string()
+    } else {
+        t.to_string()
+    }
+}
+
+/// 1-based (row, character-column) of a byte offset. Parser spans use
+/// byte columns; converting to character columns keeps the dump useful
+/// for files with multi-byte identifiers.
+fn offset_to_line_char_col(source: &str, offset: usize) -> (usize, usize) {
+    let offset = offset.min(source.len());
+    let before = &source[..offset];
+    let row = before.matches('\n').count() + 1;
+    let line_start = before.rfind('\n').map(|i| i + 1).unwrap_or(0);
+    let col = source[line_start..offset].chars().count() + 1;
+    (row, col)
+}
+
+/// Inverse of [`offset_to_line_char_col`]: byte offset of the 1-based
+/// (row, character-column) position. `None` when the row is past the
+/// last line; a column past the line end clamps to the line end.
+fn line_char_col_to_offset(source: &str, row: usize, col: usize) -> Option<usize> {
+    let mut offset = 0usize;
+    for (index, line) in source.split('\n').enumerate() {
+        if index + 1 == row {
+            let mut bytes = 0usize;
+            for ch in line.chars().take(col.saturating_sub(1)) {
+                bytes += ch.len_utf8();
+            }
+            return Some((offset + bytes).min(offset + line.len()));
+        }
+        offset += line.len() + 1;
+    }
+    None
+}
+
+/// Record the first plain assignment to each name in a scope's body.
+///
+/// R has no separate block scoping, so assignments inside `if`/`for`/
+/// `while` bodies and braced value blocks bind in the enclosing function
+/// scope. Function-literal bodies are excluded: those are their own
+/// scopes (recorded separately), though the *name* bound to a function
+/// literal (`inner <- function(...)`) is itself a local of this scope.
+fn collect_local_bindings(stmts: &[ry_core::ast::Stmt], out: &mut HashMap<String, ry_core::Span>) {
+    for statement in stmts {
+        match statement {
+            ry_core::ast::Stmt::Assign { target, value, .. } => {
+                match target {
+                    ry_core::ast::Expr::Ident { name, span }
+                    | ry_core::ast::Expr::String(name, span) => {
+                        out.entry(name.clone()).or_insert(*span);
+                    }
+                    // Indexed targets (`d$col <- v`) mutate an existing
+                    // binding rather than creating one; the base name's
+                    // plain assignment (if any) is found elsewhere. The
+                    // target's index arguments are not walked either: the
+                    // checker does not bind assignments hidden there
+                    // (`m[i <- 1L] <- v` leaves `i` unbound), and the dump
+                    // reports exactly what a `ry check` run infers.
+                    _ => {}
+                }
+                collect_local_bindings_in_expr(value, out);
+            }
+            ry_core::ast::Stmt::Expr(e) => collect_local_bindings_in_expr(e, out),
+            ry_core::ast::Stmt::If {
+                cond, then, else_, ..
+            } => {
+                collect_local_bindings_in_expr(cond, out);
+                collect_local_bindings(then, out);
+                if let Some(else_) = else_ {
+                    collect_local_bindings(else_, out);
+                }
+            }
+            ry_core::ast::Stmt::For {
+                name,
+                name_span,
+                iter,
+                body,
+                ..
+            } => {
+                out.entry(name.clone()).or_insert(*name_span);
+                collect_local_bindings_in_expr(iter, out);
+                collect_local_bindings(body, out);
+            }
+            ry_core::ast::Stmt::While { cond, body, .. } => {
+                collect_local_bindings_in_expr(cond, out);
+                collect_local_bindings(body, out);
+            }
+            // `return(e)` still binds any assignment inside `e`; a bare
+            // function definition binds no name in this scope.
+            ry_core::ast::Stmt::Return { value, .. } => {
+                if let Some(value) = value {
+                    collect_local_bindings_in_expr(value, out);
+                }
+            }
+            ry_core::ast::Stmt::FunctionDef { .. } => {}
+        }
+    }
+}
+
+/// Statement-level assignment scan through expression positions. R nests
+/// both statements (braced blocks) and side-effecting assignments
+/// (`f(x <- 1L)`, chained `a <- b <- 1L`, `if (flag <- f())`) inside
+/// arbitrary expressions, and all of them bind in the enclosing function
+/// scope. The recursion set mirrors the checker's own expression walker
+/// (`ry_checker::infer`'s `visit_expr`); literal leaves bind nothing.
+/// Function-literal bodies are excluded: those are their own scopes
+/// (recorded separately), though the *name* bound to a function literal
+/// (`inner <- function(...)`) is itself a local of this scope.
+fn collect_local_bindings_in_expr(
+    expr: &ry_core::ast::Expr,
+    out: &mut HashMap<String, ry_core::Span>,
+) {
+    match expr {
+        ry_core::ast::Expr::Call { func, args, .. } => {
+            collect_local_bindings_in_expr(func, out);
+            for argument in args {
+                collect_local_bindings_in_expr(&argument.value, out);
+            }
+        }
+        // Assignment operators in expression position bind the LHS in
+        // the current scope — the checker's `Expr::BinOp` arm does the
+        // same for `<-`/`<<-` (R's `<-` returns the value invisibly) and
+        // `%<>%` (which rebinds its LHS ident).
+        ry_core::ast::Expr::BinOp {
+            op:
+                ry_core::ast::BinOpKind::Assign
+                | ry_core::ast::BinOpKind::SuperAssign
+                | ry_core::ast::BinOpKind::PipeAssign,
+            lhs,
+            rhs,
+            ..
+        } => {
+            match lhs.as_ref() {
+                ry_core::ast::Expr::Ident { name, span }
+                | ry_core::ast::Expr::String(name, span) => {
+                    out.entry(name.clone()).or_insert(*span);
+                }
+                // Replacement-function targets (`names(d) <- v`) and
+                // indexed targets mutate; the checker's assignment arms
+                // decide those, not this walk.
+                _ => {}
+            }
+            collect_local_bindings_in_expr(lhs, out);
+            collect_local_bindings_in_expr(rhs, out);
+        }
+        ry_core::ast::Expr::BinOp { lhs, rhs, .. } => {
+            collect_local_bindings_in_expr(lhs, out);
+            collect_local_bindings_in_expr(rhs, out);
+        }
+        ry_core::ast::Expr::UnaryOp { expr, .. } => collect_local_bindings_in_expr(expr, out),
+        ry_core::ast::Expr::Index { base, args, .. } => {
+            collect_local_bindings_in_expr(base, out);
+            for argument in args {
+                collect_local_bindings_in_expr(&argument.value, out);
+            }
+        }
+        ry_core::ast::Expr::Block { body, .. } => collect_local_bindings(body, out),
+        ry_core::ast::Expr::If {
+            then, else_, cond, ..
+        } => {
+            collect_local_bindings_in_expr(cond, out);
+            collect_local_bindings_in_expr(then, out);
+            if let Some(else_) = else_ {
+                collect_local_bindings_in_expr(else_, out);
+            }
+        }
+        // Function literals are separate scopes; their bodies are indexed
+        // by index_scope_bodies instead. Literals and `Unknown` bind
+        // nothing.
+        _ => {}
+    }
+}
+
+/// Map every function body in the file to its start byte, so each
+/// recorded scope can look up its own local bindings. Mirrors the
+/// statement recursion of `collect_local_bindings` to find nested named
+/// functions at any block depth.
+fn index_scope_bodies(
+    stmts: &[ry_core::ast::Stmt],
+    index: &mut HashMap<usize, HashMap<String, ry_core::Span>>,
+) {
+    for statement in stmts {
+        match statement {
+            ry_core::ast::Stmt::Assign { value, .. } => match value {
+                ry_core::ast::Expr::Function { body, span, .. } => {
+                    index_function_body(*span, body, index);
+                }
+                _ => index_scope_bodies_in_expr(value, index),
+            },
+            ry_core::ast::Stmt::FunctionDef { body, span, .. } => {
+                index_function_body(*span, body, index);
+            }
+            ry_core::ast::Stmt::If {
+                cond, then, else_, ..
+            } => {
+                index_scope_bodies_in_expr(cond, index);
+                index_scope_bodies(then, index);
+                if let Some(else_) = else_ {
+                    index_scope_bodies(else_, index);
+                }
+            }
+            ry_core::ast::Stmt::For { iter, body, .. } => {
+                index_scope_bodies_in_expr(iter, index);
+                index_scope_bodies(body, index);
+            }
+            ry_core::ast::Stmt::While { cond, body, .. } => {
+                index_scope_bodies_in_expr(cond, index);
+                index_scope_bodies(body, index);
+            }
+            ry_core::ast::Stmt::Expr(e) => index_scope_bodies_in_expr(e, index),
+            ry_core::ast::Stmt::Return { value, .. } => {
+                if let Some(value) = value {
+                    index_scope_bodies_in_expr(value, index);
+                }
+            }
+        }
+    }
+}
+
+/// Same expression recursion as [`collect_local_bindings_in_expr`]: named
+/// functions are reachable through any expression position (a callback's
+/// body, a call argument, an index argument), and only finding them there
+/// gives their scopes `function_locals` entries.
+fn index_scope_bodies_in_expr(
+    expr: &ry_core::ast::Expr,
+    index: &mut HashMap<usize, HashMap<String, ry_core::Span>>,
+) {
+    match expr {
+        ry_core::ast::Expr::Call { func, args, .. } => {
+            index_scope_bodies_in_expr(func, index);
+            for argument in args {
+                index_scope_bodies_in_expr(&argument.value, index);
+            }
+        }
+        ry_core::ast::Expr::BinOp { lhs, rhs, .. } => {
+            index_scope_bodies_in_expr(lhs, index);
+            index_scope_bodies_in_expr(rhs, index);
+        }
+        ry_core::ast::Expr::UnaryOp { expr, .. } => index_scope_bodies_in_expr(expr, index),
+        ry_core::ast::Expr::Index { base, args, .. } => {
+            index_scope_bodies_in_expr(base, index);
+            for argument in args {
+                index_scope_bodies_in_expr(&argument.value, index);
+            }
+        }
+        ry_core::ast::Expr::Block { body, .. } => index_scope_bodies(body, index),
+        ry_core::ast::Expr::If {
+            then, else_, cond, ..
+        } => {
+            index_scope_bodies_in_expr(cond, index);
+            index_scope_bodies_in_expr(then, index);
+            if let Some(else_) = else_ {
+                index_scope_bodies_in_expr(else_, index);
+            }
+        }
+        // An anonymous function literal gets no `ScopeRecord` of its own
+        // (the checker infers it in discarding mode), but named functions
+        // defined inside it complete and are recorded — walk the body so
+        // those nested definitions are indexed. The literal itself needs
+        // no `function_locals` entry: no record will ever look it up.
+        ry_core::ast::Expr::Function { body, .. } => index_scope_bodies(body, index),
+        _ => {}
+    }
+}
+
+fn index_function_body(
+    span: ry_core::Span,
+    body: &[ry_core::ast::Stmt],
+    index: &mut HashMap<usize, HashMap<String, ry_core::Span>>,
+) {
+    let mut locals = HashMap::new();
+    collect_local_bindings(body, &mut locals);
+    index.insert(span.start, locals);
+    index_scope_bodies(body, index);
+}
+
+/// Per-scope classification inputs derived once per file.
+struct ScopeInfo<'a> {
+    record: &'a ry_checker::ScopeRecord,
+    params: HashMap<&'a str, ry_core::Span>,
+    locals: HashMap<&'a str, ry_core::Span>,
+}
+
+/// Turn one file's scope records into the JSON dump shape.
+///
+/// Binding kinds (documented in the README section for `dump-types`):
+/// - `param`: still marked as a formal in the recorded scope (an
+///   overwritten formal degrades to `local`, matching R's rebinding).
+/// - `local`: first assigned inside this scope's own body.
+/// - `closed-over`: function scopes only — present because the body's
+///   scope was cloned from the enclosing one at definition time.
+/// - `imported`: top-level bindings the file never assigns (ambient names
+///   supplied by the host environment, e.g. Shiny server fragments).
+fn assemble_file_dump(
+    path: &str,
+    file: &ry_core::SourceFile,
+    records: Vec<ry_checker::ScopeRecord>,
+    positions: &[(usize, usize)],
+) -> FileDump {
+    // Sort by start position and drop duplicates (an injected-expression
+    // re-walk can complete the same literal twice). The dedup key
+    // includes the kind: a leading function literal's span starts at the
+    // same byte 0 as the whole-file top scope, and keying on the offset
+    // alone would drop that top scope and every top-level binding with
+    // it.
+    let mut records = records;
+    records.sort_by_key(|record| record.span.start);
+    records.dedup_by(|a, b| a.kind == b.kind && a.span.start == b.span.start);
+
+    let mut function_locals: HashMap<usize, HashMap<String, ry_core::Span>> = HashMap::new();
+    index_scope_bodies(&file.stmts, &mut function_locals);
+    let mut top_locals = HashMap::new();
+    collect_local_bindings(&file.stmts, &mut top_locals);
+
+    let infos: Vec<ScopeInfo> = records
+        .iter()
+        .map(|record| ScopeInfo {
+            record,
+            params: record
+                .params
+                .iter()
+                .map(|(name, span)| (name.as_str(), *span))
+                .collect(),
+            locals: match record.kind {
+                ry_checker::ScopeRecordKind::Function => function_locals
+                    .get(&record.span.start)
+                    .map(|locals| {
+                        locals
+                            .iter()
+                            .map(|(name, span)| (name.as_str(), *span))
+                            .collect()
+                    })
+                    .unwrap_or_default(),
+                ry_checker::ScopeRecordKind::Top => top_locals
+                    .iter()
+                    .map(|(name, span)| (name.as_str(), *span))
+                    .collect(),
+            },
+        })
+        .collect();
+
+    // Enclosing-scope chains are shared by every binding of a scope, so
+    // compute them once per file instead of once per closed-over lookup.
+    let chains = enclosing_scope_chains(&infos);
+
+    // Which scopes does --position select? Without positions, all. With
+    // them, the innermost containing scope for each position (the union,
+    // deduplicated). Byte offsets make containment exact regardless of
+    // encoding. One pass records both the selected scopes and, per
+    // scope, the offsets that selected it — the latter drives
+    // binding-visibility filtering below.
+    let mut selected: Vec<usize> = Vec::new();
+    let mut selecting_offsets: HashMap<usize, Vec<usize>> = HashMap::new();
+    if positions.is_empty() {
+        selected.extend(0..infos.len());
+    } else {
+        for &(row, col) in positions {
+            let Some(offset) = line_char_col_to_offset(&file.source, row, col) else {
+                continue;
+            };
+            let mut best: Option<(usize, usize)> = None;
+            for (index, info) in infos.iter().enumerate() {
+                let span = info.record.span;
+                if span.start <= offset && offset < span.end {
+                    let extent = span.end - span.start;
+                    if best.is_none_or(|(extent_so_far, _)| extent < extent_so_far) {
+                        best = Some((extent, index));
+                    }
+                }
+            }
+            if let Some((_, index)) = best {
+                selecting_offsets.entry(index).or_default().push(offset);
+            }
+        }
+        selected.extend(selecting_offsets.keys().copied());
+        selected.sort_unstable();
+        selected.dedup();
+    }
+
+    let scopes = selected
+        .into_iter()
+        .map(|index| {
+            let info = &infos[index];
+            let record = info.record;
+            let source = &file.source;
+
+            let mut bindings: Vec<BindingDump> = record
+                .scope
+                .bindings
+                .iter()
+                .map(|(name, ty)| {
+                    // A formal counts as `param` only in the scope that
+                    // declares it: scope snapshots are cloned from the
+                    // enclosing function, which also clones the
+                    // parameter-marker set, so a captured outer formal
+                    // must not be reclassified. A declared formal whose
+                    // marker is gone was reassigned in the body (R
+                    // rebinds rather than narrows), so it degrades to
+                    // `local` at its reassignment site.
+                    let is_formal_here = info.params.contains_key(name.as_str());
+                    let is_param = is_formal_here && record.scope.parameter_bindings.contains(name);
+                    let local_span = info.locals.get(name.as_str()).copied();
+                    let (kind, definition_span) = if is_param {
+                        (
+                            "param",
+                            info.params.get(name.as_str()).copied().or(local_span),
+                        )
+                    } else if let Some(span) = local_span {
+                        ("local", Some(span))
+                    } else if record.kind == ry_checker::ScopeRecordKind::Function {
+                        ("closed-over", None)
+                    } else {
+                        ("imported", None)
+                    };
+                    // A closed-over binding has no site in this scope;
+                    // point at the definition in the nearest enclosing
+                    // scope that binds the name, when one was recorded.
+                    let definition_span = definition_span
+                        .or_else(|| enclosing_binding_span(&infos, &chains[index], name.as_str()));
+                    (name, ty, kind, definition_span, is_formal_here)
+                })
+                .filter(|(_, _, _, definition_span, is_formal_here)| {
+                    // Visibility at the selecting positions: formals bind
+                    // at call entry (always visible); a local assigned
+                    // after every selecting position is not yet in scope
+                    // there; a binding with no resolvable site is kept
+                    // (it predates the body).
+                    let Some(offsets) = selecting_offsets.get(&index) else {
+                        return true;
+                    };
+                    if *is_formal_here {
+                        return true;
+                    }
+                    let Some(span) = definition_span else {
+                        return true;
+                    };
+                    offsets.iter().any(|offset| span.start <= *offset)
+                })
+                .map(|(name, ty, kind, definition_span, _)| BindingDump {
+                    name: name.clone(),
+                    kind,
+                    type_: dump_type_string(ty),
+                    start: definition_span.map(|span| offset_to_line_char_col(source, span.start)),
+                })
+                .collect();
+            bindings.sort_by(|a, b| a.name.cmp(&b.name));
+
+            ScopeDump {
+                kind: match record.kind {
+                    ry_checker::ScopeRecordKind::Function => "function",
+                    ry_checker::ScopeRecordKind::Top => "top",
+                },
+                name: record.name.clone(),
+                start: offset_to_line_char_col(source, record.span.start),
+                end: offset_to_line_char_col(source, record.span.end),
+                bindings,
+            }
+        })
+        .collect();
+
+    FileDump {
+        path: path.to_string(),
+        scopes,
+    }
+}
+
+/// Sorted (nearest-first) enclosing-scope index chain for every scope:
+/// each entry lists the other scopes whose span contains it, ordered by
+/// smallest extent first. Computed once per file so every binding lookup
+/// in a scope reuses the same chain.
+fn enclosing_scope_chains(infos: &[ScopeInfo]) -> Vec<Vec<usize>> {
+    (0..infos.len())
+        .map(|index| {
+            let inner = infos[index].record.span;
+            let mut chain: Vec<usize> = (0..infos.len())
+                .filter(|other| {
+                    *other != index
+                        && infos[*other].record.span.start <= inner.start
+                        && inner.end <= infos[*other].record.span.end
+                })
+                .collect();
+            // Nearest first: smallest enclosing extent.
+            chain.sort_by_key(|other| {
+                infos[*other].record.span.end - infos[*other].record.span.start
+            });
+            chain
+        })
+        .collect()
+}
+
+/// Definition site of `name` in the nearest recorded scope enclosing the
+/// scope that owns `chain`, walking outward. Used for closed-over
+/// bindings, whose only site in this file lives in an outer scope.
+fn enclosing_binding_span(
+    infos: &[ScopeInfo],
+    chain: &[usize],
+    name: &str,
+) -> Option<ry_core::Span> {
+    for index in chain {
+        let info = &infos[*index];
+        if let Some(span) = info.params.get(name).copied() {
+            return Some(span);
+        }
+        if let Some(span) = info.locals.get(name).copied() {
+            return Some(span);
+        }
+    }
+    None
+}
+
+pub(crate) fn run_dump_types(
+    files: Vec<PathBuf>,
+    project_root: Option<PathBuf>,
+    format: &str,
+    positions: Vec<(usize, usize)>,
+) -> Result<ExitCode> {
+    if format != "json" {
+        return Err(miette::miette!(
+            "unknown --format `{}`; expected one of: json",
+            format
+        ));
+    }
+
+    // Config discovery mirrors `ry check`: anchored at the first input,
+    // missing config is fine, malformed config aborts. The config's
+    // directory is kept as `config_root` — it anchors the `exclude`
+    // patterns below and is the resolution-root fallback for non-package
+    // files, exactly as in `run_check`.
+    let search_start = files.first().cloned().unwrap_or_else(|| PathBuf::from("."));
+    let (config_root, cfg) = match config::Config::discover(&search_start) {
+        Ok(Some((path, cfg))) => (path.parent().map(PathBuf::from), cfg),
+        Ok(None) => (None, config::Config::defaults()),
+        Err(e) => {
+            eprintln!("ry: {}", e);
+            return Ok(ExitCode::FAILURE);
+        }
+    };
+
+    let mut all_paths = Vec::new();
+    for root in &files {
+        if !root.exists() {
+            eprintln!("ry: {}: no such file or directory", root.display());
+            return Ok(ExitCode::FAILURE);
+        }
+        let result = ry_workspace::discover_r_files(
+            root,
+            config_root.as_deref(),
+            &cfg,
+            cfg.check_test_fixtures,
+        );
+        all_paths.extend(result.files);
+        report_truncation(&result.truncated, root);
+    }
+    sort_and_deduplicate_paths(&mut all_paths);
+
+    // Parallel parsing with the same thread-local parser pool as
+    // `ry check` (tree-sitter parsers are not Send).
+    thread_local! {
+        static DUMP_PARSER: std::cell::RefCell<Option<ry_core::RParser>> =
+            const { std::cell::RefCell::new(None) };
+    }
+    use rayon::prelude::*;
+    // Ok(Some(..)) = parsed; Ok(None) = unparseable (warned and skipped;
+    // the dump still covers every parseable file); Err(..) = unreadable
+    // (fatal, reported after the join).
+    type ParseOutcome = Result<Option<(String, String, ry_core::SourceFile)>, String>;
+    let results: Vec<ParseOutcome> = all_paths
+        .par_iter()
+        .map(|path| {
+            let src = match read_r_source(path) {
+                Ok(src) => src,
+                Err(e) => return Err(format!("{}: {}", path.display(), e)),
+            };
+            let path_str = path.to_string_lossy().to_string();
+            let file = DUMP_PARSER.with(|cell| {
+                let mut slot = cell.borrow_mut();
+                let parser = slot.get_or_insert_with(|| {
+                    ry_core::RParser::new().expect("parser init (thread-local)")
+                });
+                parser.parse(&path_str, &src)
+            });
+            match file {
+                Ok(file) => Ok(Some((path_str, src, file))),
+                Err(e) => {
+                    eprintln!(
+                        "ry: {}: parse error: {e}; file omitted from dump",
+                        path.display()
+                    );
+                    Ok(None)
+                }
+            }
+        })
+        .collect();
+
+    let mut parsed: Vec<(String, String, ry_core::SourceFile)> = Vec::new();
+    for result in results {
+        match result {
+            Err(read_error) => {
+                eprintln!("ry: {read_error}");
+                return Ok(ExitCode::FAILURE);
+            }
+            Ok(Some(entry)) => parsed.push(entry),
+            Ok(None) => {}
+        }
+    }
+
+    let user_stubs = load_user_stubs(&cfg.typeshed);
+
+    // Same per-package grouping as `ry check`: each DESCRIPTION root is
+    // its own library namespace. Non-package scripts share one group
+    // rooted at --project-root, else the config root (the directory
+    // owning the discovered ry.toml), else the working directory —
+    // `run_check_once`'s fallback chain with --project-root overriding.
+    let mut groups: std::collections::BTreeMap<Option<PathBuf>, Vec<usize>> =
+        std::collections::BTreeMap::new();
+    for (index, (path, _, _)) in parsed.iter().enumerate() {
+        groups
+            .entry(enclosing_package_root(std::path::Path::new(path)))
+            .or_default()
+            .push(index);
+    }
+
+    let mut records_by_path: HashMap<String, Vec<ry_checker::ScopeRecord>> = HashMap::new();
+    for (group_root, indices) in &groups {
+        let resolution_root = group_root
+            .clone()
+            .or_else(|| project_root.clone())
+            .or_else(|| config_root.clone())
+            .unwrap_or_else(|| std::env::current_dir().unwrap_or_else(|_| PathBuf::from(".")));
+        let package_scope = ry_workspace::resolve_workspace_context(
+            &resolution_root,
+            &cfg,
+            ry_workspace::ResolutionEnvironment {
+                files: indices.iter().map(|index| &parsed[*index].2).collect(),
+                user_stubs: &user_stubs,
+            },
+        )
+        .map_err(|error| miette::miette!(error))?;
+        let analysis_files: Vec<_> = indices
+            .iter()
+            .map(|index| {
+                let (path, _, file) = &parsed[*index];
+                (path.clone(), 0, std::sync::Arc::new(file.clone()))
+            })
+            .collect();
+        let check_input = check::CheckInput {
+            files: analysis_files,
+            user_stubs: std::sync::Arc::clone(&user_stubs),
+            workspace: Some(ry_workspace::WorkspaceContext {
+                attached_packages: package_scope.attached_packages,
+                bare_bindings: package_scope.bare_bindings,
+                external_bindings: package_scope.external_bindings,
+                imported_bindings: package_scope.imported_bindings,
+                s3_methods: package_scope.s3_methods,
+                load_bindings: package_scope.load_bindings,
+                degraded_scopes: Vec::new(),
+            }),
+        };
+        // `ry check` prints one summary line per degraded scope; keep the
+        // note on stderr here too so a dump over the same project reports
+        // the same precision loss without polluting the JSON on stdout.
+        for (path, reason) in &package_scope.degraded_scopes {
+            eprintln!(
+                "ry: {}: degraded scope ({reason}); serialized data file(s) over the byte cap fell back to file stems",
+                path.display()
+            );
+        }
+        for (path, records) in check::check_project_with_scope_capture(check_input) {
+            records_by_path.insert(path, records);
+        }
+    }
+
+    let dump = TypesDump {
+        files: parsed
+            .iter()
+            .map(|(path, _src, file)| {
+                let records = records_by_path.remove(path).unwrap_or_default();
+                assemble_file_dump(path, file, records, &positions)
+            })
+            .collect(),
+    };
+    println!("{}", serde_json::to_string_pretty(&dump).into_diagnostic()?);
+    // Diagnostics (if any) never affect the dump's exit code.
+    Ok(ExitCode::SUCCESS)
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn dump_position_parser_accepts_pairs_and_rejects_garbage() {
+        assert_eq!(parse_dump_position("3:14").unwrap(), (3, 14));
+        assert_eq!(parse_dump_position(" 12 : 1 ").unwrap(), (12, 1));
+        assert!(parse_dump_position("3").is_err());
+        assert!(parse_dump_position("a:b").is_err());
+        assert!(parse_dump_position("0:1").is_err(), "rows are 1-based");
+        assert!(parse_dump_position("1:0").is_err(), "cols are 1-based");
+    }
+
+    #[test]
+    fn dump_position_offsets_round_trip_and_clamp() {
+        // Multi-byte characters: columns must count characters, not bytes.
+        let src = "a <- 1L\n#\u{e9} <- 2L\nlast <- 3L\n";
+        assert_eq!(offset_to_line_char_col(src, 0), (1, 1));
+        // Start of line 3.
+        assert_eq!(
+            offset_to_line_char_col(src, src.find("last").unwrap()),
+            (3, 1)
+        );
+        // The identifier on line 2 starts after `#`, a multi-byte char.
+        let ident = src.find("<- 2L").unwrap();
+        let (row, col) = offset_to_line_char_col(src, ident);
+        assert_eq!((row, col), (2, 4));
+        assert_eq!(
+            line_char_col_to_offset(src, row, col),
+            Some(ident),
+            "round trip"
+        );
+        // Row past the end matches nothing; column past the line end
+        // clamps to the line end.
+        assert_eq!(line_char_col_to_offset(src, 99, 1), None);
+        assert_eq!(
+            line_char_col_to_offset(src, 1, 500),
+            Some(src.find('\n').unwrap())
+        );
+    }
+
+    #[test]
+    fn dump_type_string_renders_unknown_and_display_forms() {
+        assert_eq!(dump_type_string(&ry_core::RType::unknown()), "unknown");
+        let integer =
+            ry_core::RType::new(ry_core::types::Mode::Integer, ry_core::types::Length::One);
+        assert_eq!(dump_type_string(&integer), "integer<len=1>");
+    }
+
+    #[test]
+    fn dump_local_binding_scan_skips_nested_function_bodies() {
+        let mut parser = ry_core::RParser::new().unwrap();
+        let file = parser
+            .parse(
+                "a.R",
+                "outer <- function(x) {\n  keep <- 1L\n  inner <- function(y) { skip <- 2L }\n  if (x) keep2 <- 3L\n  for (i in 1:3) keep3 <- i\n  keep\n}\n",
+            )
+            .unwrap();
+        // The function body's own locals, extracted from the statement
+        // tree: nested `skip` belongs to the inner scope, not this one.
+        let ry_core::ast::Stmt::Assign { value, .. } = &file.stmts[0] else {
+            panic!("expected assignment");
+        };
+        let ry_core::ast::Expr::Function { body, .. } = value else {
+            panic!("expected function literal");
+        };
+        let mut locals = std::collections::HashMap::new();
+        collect_local_bindings(body, &mut locals);
+        let mut names: Vec<_> = locals.keys().map(String::as_str).collect();
+        names.sort_unstable();
+        assert_eq!(names, vec!["i", "inner", "keep", "keep2", "keep3"]);
+    }
+}
